Relaxe eslint config
Allow for..of, allow redundant else statements
for style purposes.

Forbids multiple statements per line. This specifically
prevents the following construction which is equivalent
to using if..else if, but unusual to read and may introduce
errors when refactoring the code not to use return statements.
By allowing redundant else statements we can just use an
actual if..else if construction.

    if (foo) {
      return 42;
    } if (bar) {
      return 23;
    }